Also expose the game points so far.
This changes the format of the /scores/league endpoint such that the
previous dictionary of league scores is now under a 'league_points'
key, and a new 'game_points' key has been added with a similar dictionary
of total game points. The collection of game points is deliberately
limited to being summed over the league.