Add endpoint to get arbitrary match sequences
This patch adds an endpoint at /matches/<arena> that takes an argument
in its query string called "numbers".  "numbers" is a comma-separated
list of match numbers that the requester would like information about.

This endpoint also supports three special values that can be used in
place of match numbers:

 - "current": for the current match
 - "next": for the next match
 - "next+1": for the next-but-one match

Note that there may be no current match, and there may be no next
match.  So those items may resolve to null.

For example: /matches/A?numbers=current,next

May return { matches = [ null, { ... match info ... } ] } if there is
no current match, but there is on in the future.

(Maybe in the future "next+$N" will be supported, but for now it's
not. We have an application that needs next+1 now.)