Add a /state endpoint.
This will emit a short description of the current state, specifically
the git revision. This is useful so we can see which revision is
actually being used, and also potentially to reduce the http polling
that occurs.

Update test setup to include fudging the path of the gitdir in the
temporary copy of the compstate repo used.

Change-Id: I0b5afc29c1f8dc09cc79be16977932a453a27d33