Add a compstate update mechanism.
During normal operation, the http backend will (at most once each
5 seconds) look for a specific file in the compstate repo. If this
file is present, nothing happens (ie it uses the data it already has).
If the file is missing, then it will create and lock the file and
then re-load the SRComp instance.

Updates to the compstate should be done using the provided 'update'
script. This will fetch the latest revisions from the 'origin' remote,
acquire a lock on the same file and then checkout origin/master.
Once complete, the script deletes the file.

As the http process will only re-load the compstate after the file
has been removed, and the file only gets removed when the update is
complete, it may appear that the locks aren't needed. However, there
are two cases where the file is not present:
  * initial startup
  * after an update
During these times, there are potential race conditions between the
data being loaded and it being updated. The locks ensure that even
here, access to the state is restricted and suitably ordered.

This is related to #2363 (Build updating of competition state), but
doesn't resolve it.

Change-Id: Ibee155a571b6f62d5470bffade3b0c4df95a0f15