Add logging capability
This copies the logging configs (and escaping formatter) from nemesis,
though uses a slightly different setup mechanism as here we don't
have any existing local config of note.

For the moment, this just adds simple logging when we re-load the
compstate repo, but now that the emitters are set up it's pretty
easy to add more logging points.

Change-Id: Iefb0d8e77cfe08041617e7eca697f55fbb0781e6