Remove duplicate use of the term 'period'
This originally meant a stretch of time during which matches would
be run, but appears to have also been used to mean the duration of
a match. Since then the latter appears to have been epanded upon,
causing confusion. Replace the with 'duration' where suitable or
with 'slot' when refering to the specific timeslot a match occurs
within.

Change-Id: I207ec7c2ca9e93196b9571c5e67223a3bea7544e