Expose the details of the match periods
Add a '/periods' endpoint which contains details of each period, but
not the contents of all the matches they contain. Instead, just
provide the first and last match ids which each period contains so
that the caller can look that up separately.

Patch set #2: rebased, add/update tests, remove 'matches' key when
              there aren't any.

Change-Id: Ie5c3f7a09bd9d934cbba8dce5f694bbb45040e6c