Compensate for an apparent bug in freezegun
Under Python 2, freezegun causes datetime.now() to report the same
actual time both with and without a timezone. That is, the numeric
time changes, but only because the proper offset has been applied.

Under Python 3, freezegun causes datetime.now() to report different
times when a timezone is applied. Specifically it reports the same
numeric time, ignoring the offset which should be applied from the
timezone value.

We therefore need to compensate by adjusting the apparently frozen
time under Python 3 such that we've already applied the proper offset.

Change-Id: Ib18d0a61911d03595369667d1284c08846e14b14