Switch to using a RawCompstate to make the changes
This removes the need to do process handling locally, and means we
can rely on the existing class's better wrapping of git functions.

Specifically, it turns out that rev-parse will appear to succeed for
inputs which look like a commit, but aren't actually in the repo.
This is fixed by appending a tag to the commit id, which the
RawCompstate class already does.

Change-Id: Ic247dd09567f7875d96c4f20fd1dd67f0df62428