Revert timezone issues workaround
This logically reverts 63d01cbd0b09017a70b0e770939d116085d6b5cd,
which turned out not to be a bug in freezegun but in fact in Python
itself (https://bugs.python.org/issue23600). Since that bug is now
fixed in all the latest 3.x releases we support there is no need
for the workaround (and in fact it causes the tests to fail).

Change-Id: I10db677c51e1a47a40ae3f67b5c6e2f60b219f04